requirements.txt
streamlit
pandas
numpy
gspread
google-auth
gspread-dataframe
python-dotenv
geopy
folium
streamlit-folium